pyproject.toml
[tool.poetry]
name = "shopapi"
version = "0.1.0"
description = "REST API for basic e-commerce platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
mysqlclient = "^2.1.1"
django-phonenumber-field = "^7.0.2"
django-countries = "^7.5.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
